pyproject.toml
[project]
name = "perplexity-clone-app"
version = "0.1.0"
description = "FastAPI backend application for a Perplexity AI clone. Supports user authentication, question answering from various sources, real-time web search and crawling and session management."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "authlib>=1.3.0",
    "bcrypt>=4.1.0",
    "boto3>=1.40.74",
    "crawl4ai>=0.7.7",
    "ddgs>=9.9.1",
    "fastapi>=0.121.1",
    "firecrawl-py>=4.8.0",
    "google-api-python-client>=2.187.0",
    "google-search-results>=2.4.2",
    "jinja2>=3.1.6",
    "langchain>=1.0.5",
    "langchain-anthropic>=1.1.0",
    "langchain-astradb>=1.0.0",
    "langchain-cohere>=0.5.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-exa>=1.0.0",
    "langchain-google-genai>=3.0.3",
    "langchain-groq>=1.0.1",
    "langchain-huggingface>=1.0.1",
    "langchain-mistralai>=1.0.1",
    "langchain-pinecone>=0.2.13",
    "langchain-postgres>=0.0.16",
    "langchain-tavily>=0.2.13",
    "langgraph>=1.0.2",
    "langgraph-checkpoint-postgres>=3.0.1",
    "langgraph-checkpoint-redis>=0.2.1",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "langsmith>=0.4.42",
    "logfire[fastapi,httpx,requests,sqlalchemy,system-metrics]>=4.14.2",
    "loguru>=0.7.3",
    "pinecone>=7.3.0",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary,pool]>=3.2.12",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.3.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "upstash-redis>=1.5.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "fastapi[standard]>=0.121.1",
    "ipykernel>=7.1.0",
    "pip-system-certs>=5.3",
]
